--- plugins/transforms/googlesheet/src/main/doc/google-sheets-input.adoc
@@ -1,36 +1,0 @@
-////
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-////
-:documentationPath: /plugins/transforms/
-:language: en_US
-:page-alternativeEditUrl: https://github.com/apache/incubator-hop/edit/master/plugins/transforms/googleanalytics/src/main/doc/google-sheets-input.adoc
-
-= Google Sheets Input
-
-== Service Account Tab
-
-Lets you pick your Google service account client secret json file. The `Test` button Lets you test your access to the API. On success, you should see a success message.
-
-== Spreadsheet tab
-
-This lets you specify or browse for spreadsheets existing in the service account drive or for the ones that are shared with the service account email.
-
-== Fields tab
-
-Lets you select from the fields of the sheet. Fields name are always defined in the first line of the Google spreadsheet.
-
-`Get Fields` lets you get fields and guess their types, format, precision, decimal and group delimiter as well as trim type.
-
