--- plugins/transforms/googlesheet/src/main/resources/org/apache/hop/pipeline/transforms/googlesheet/messages/messages_en_FR.properties
@@ -1,57 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-#
-
-GoogleSheetsPluginInputDialog.TransformName.Label=HOP Google Sheet Input
-GoogleSheetsPluginInputDialog.Shell.Title=Google Sheet Input V4
-GoogleSheetsPluginInput.Log.LineNumber=Linenr
-GoogleSheetsPluginInput.transform.Name=HOP Google Sheets Input
-GoogleSheetsPluginInput.injection.SHEET=Specs for inbound Spreadsheet 
-GoogleSheetsPluginInput.injection.spreadsheetKey=ID of the Google Spreadsheet
-GoogleSheetsPluginInput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
-GoogleSheetsPluginInput.injection.INPUT_FIELDS=Name of the inbound fields
-GoogleSheetsPluginInput.injection.INPUT_NAME=Field Name
-GoogleSheetsPluginInput.injection.INPUT_POSITION=Position
-GoogleSheetsPluginInput.injection.INPUT_LENGTH=Length
-GoogleSheetsPluginInput.injection.INPUT_IGNORE=Ignore
-GoogleSheetsPluginInput.injection.INPUT_FORMAT=Format
-GoogleSheetsPluginInput.injection.INPUT_PRECISION=Precision
-GoogleSheetsPluginInput.injection.INPUT_CURRENCY=Currency
-GoogleSheetsPluginInput.injection.INPUT_DECIMAL=Decimal
-GoogleSheetsPluginInput.injection.INPUT_GROUP=Group
-GoogleSheetsPluginInput.injection.INPUT_REPEAT=Repeat
-GoogleSheetsPluginInput.injection.INPUT_NULL_STRING=Null String
-GoogleSheetsPluginInput.injection.INPUT_IF_NULL=If Null
-GoogleSheetsPluginInput.injection.jsonCrendentialPath=Path to the Google JSON credential file
-
-
-GoogleSheetsPluginOutputDialog.transform.Name.Label=HOP Google Sheet Output
-GoogleSheetsPluginOutputDialog.Shell.Title=Google Sheet Output V4
-GoogleSheetsPluginOutput.Log.LineNumber=Linenr
-GoogleSheetsPluginOutput.transform.Name=HOP Google Sheets Output
-GoogleSheetsPluginOutputDialog.Create.Label=Création de fivhier si inexistant
-GoogleSheetsPluginOutputDialog.Append.Label=Ajout en fin de sheet
-GoogleSheetsPluginOutputDialog.Share.Label=Partage des droits (RW) avec
-GoogleSheetsPluginOutputDialog.Share.LabelDW=Permission de domaine :
-GoogleSheetsPluginOutput.injection.SHEET=Specification pour la sheet crée 
-GoogleSheetsPluginOutput.injection.spreadsheetKey=ID de la Google Spreadsheet
-GoogleSheetsPluginOutput.injection.worksheetId=Nom de l'onglet dans lequel écrire
-GoogleSheetsPluginOutput.injection.Create=Créer une sheet si elle n'existe pas
-GoogleSheetsPluginOutput.injection.Append=Ajout à la sheet existante (sans header)
-GoogleSheetsPluginOutput.injection.Email=Partager la sheet avec un user
-GoogleSheetsPluginOutput.injection.Domain=Permission de domaine 
-GoogleSheetsPluginOutput.injection.jsonCrendentialPath=Chemin vers Google JSON credential file
