--- plugins/transforms/googlesheet/src/main/resources/org/apache/hop/pipeline/transforms/googlesheet/messages/messages_en_US.properties
@@ -1,57 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-#
-
-GoogleSheetsPluginInputDialog.TransformName.Label=Google Sheet Input
-GoogleSheetsPluginInputDialog.Shell.Title=Google Sheet Input V4
-GoogleSheetsPluginInput.Log.LineNumber=Linenr
-GoogleSheetsPluginInput.transform.Name=Google Sheets Input
-GoogleSheetsPluginInput.injection.SHEET=Specs for inbound Spreadsheet 
-GoogleSheetsPluginInput.injection.spreadsheetKey=ID of the Google Spreadsheet
-GoogleSheetsPluginInput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
-GoogleSheetsPluginInput.injection.INPUT_FIELDS=Name of the inbound fields
-GoogleSheetsPluginInput.injection.INPUT_NAME=Field Name
-GoogleSheetsPluginInput.injection.INPUT_POSITION=Position
-GoogleSheetsPluginInput.injection.INPUT_LENGTH=Length
-GoogleSheetsPluginInput.injection.INPUT_IGNORE=Ignore
-GoogleSheetsPluginInput.injection.INPUT_FORMAT=Format
-GoogleSheetsPluginInput.injection.INPUT_PRECISION=Precision
-GoogleSheetsPluginInput.injection.INPUT_CURRENCY=Currency
-GoogleSheetsPluginInput.injection.INPUT_DECIMAL=Decimal
-GoogleSheetsPluginInput.injection.INPUT_GROUP=Group
-GoogleSheetsPluginInput.injection.INPUT_REPEAT=Repeat
-GoogleSheetsPluginInput.injection.INPUT_NULL_STRING=Null String
-GoogleSheetsPluginInput.injection.INPUT_IF_NULL=If Null
-GoogleSheetsPluginInput.injection.jsonCrendentialPath=Path to the Google JSON credential file
-
-
-GoogleSheetsPluginOutputDialog.TransformName.Label=Google Sheet Output
-GoogleSheetsPluginOutputDialog.Shell.Title=Google Sheet Output V4
-GoogleSheetsPluginOutput.Log.LineNumber=Linenr
-GoogleSheetsPluginOutput.transform.Name=Google Sheets Output
-GoogleSheetsPluginOutputDialog.Create.Label=Create new sheet if it does not exist
-GoogleSheetsPluginOutputDialog.Append.Label=Append to sheet
-GoogleSheetsPluginOutputDialog.Share.Label=Share Full rights (RW) with
-GoogleSheetsPluginOutputDialog.Share.LabelDW=Domain Wise Permission :
-GoogleSheetsPluginOutput.injection.SHEET=Specs for outbound Spreadsheet 
-GoogleSheetsPluginOutput.injection.spreadsheetKey=ID of the Google Spreadsheet
-GoogleSheetsPluginOutput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
-GoogleSheetsPluginOutput.injection.Create=Create Sheet if it does not exist
-GoogleSheetsPluginOutput.injection.Append=Append data to existing Sheet (without header)
-GoogleSheetsPluginOutput.injection.Email=Share the created sheet with a user
-GoogleSheetsPluginOutput.injection.Domain=Domain name for domain extenstion
-GoogleSheetsPluginOutput.injection.jsonCrendentialPath=Path to the Google JSON credential file
