+++ plugins/transforms/googlesheets/src/main/resources/org/apache/hop/pipeline/transforms/googlesheets/messages/messages_en_FR.properties
@@ -1,0 +1,57 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+
+GoogleSheetsInputDialog.TransformName.Label=Google Sheet Input
+GoogleSheetsInputDialog.Shell.Title=Google Sheet Input V4
+GoogleSheetsInput.Log.LineNumber=Linenr
+GoogleSheetsInput.transform.Name=Google Sheets Input
+GoogleSheetsInput.injection.SHEET=Specs for inbound Spreadsheet 
+GoogleSheetsInput.injection.spreadsheetKey=ID of the Google Spreadsheet
+GoogleSheetsInput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
+GoogleSheetsInput.injection.INPUT_FIELDS=Name of the inbound fields
+GoogleSheetsInput.injection.INPUT_NAME=Field Name
+GoogleSheetsInput.injection.INPUT_POSITION=Position
+GoogleSheetsInput.injection.INPUT_LENGTH=Length
+GoogleSheetsInput.injection.INPUT_IGNORE=Ignore
+GoogleSheetsInput.injection.INPUT_FORMAT=Format
+GoogleSheetsInput.injection.INPUT_PRECISION=Precision
+GoogleSheetsInput.injection.INPUT_CURRENCY=Currency
+GoogleSheetsInput.injection.INPUT_DECIMAL=Decimal
+GoogleSheetsInput.injection.INPUT_GROUP=Group
+GoogleSheetsInput.injection.INPUT_REPEAT=Repeat
+GoogleSheetsInput.injection.INPUT_NULL_STRING=Null String
+GoogleSheetsInput.injection.INPUT_IF_NULL=If Null
+GoogleSheetsInput.injection.jsonCrendentialPath=Path to the Google JSON credential file
+
+
+GoogleSheetsOutputDialog.transform.Name.Label=Google Sheet Output
+GoogleSheetsOutputDialog.Shell.Title=Google Sheet Output V4
+GoogleSheetsOutput.Log.LineNumber=Linenr
+GoogleSheetsOutput.transform.Name=Google Sheets Output
+GoogleSheetsOutputDialog.Create.Label=Création de fivhier si inexistant
+GoogleSheetsOutputDialog.Append.Label=Ajout en fin de sheet
+GoogleSheetsOutputDialog.Share.Label=Partage des droits (RW) avec
+GoogleSheetsOutputDialog.Share.LabelDW=Permission de domaine :
+GoogleSheetsOutput.injection.SHEET=Specification pour la sheet crée 
+GoogleSheetsOutput.injection.spreadsheetKey=ID de la Google Spreadsheet
+GoogleSheetsOutput.injection.worksheetId=Nom de l'onglet dans lequel écrire
+GoogleSheetsOutput.injection.Create=Créer une sheet si elle n'existe pas
+GoogleSheetsOutput.injection.Append=Ajout à la sheet existante (sans header)
+GoogleSheetsOutput.injection.Email=Partager la sheet avec un user
+GoogleSheetsOutput.injection.Domain=Permission de domaine 
+GoogleSheetsOutput.injection.jsonCrendentialPath=Chemin vers Google JSON credential file
