+++ plugins/transforms/googlesheets/src/main/resources/org/apache/hop/pipeline/transforms/googlesheets/messages/messages_en_US.properties
@@ -1,0 +1,64 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+
+GoogleSheetsInputDialog.TransformName.Label=Google Sheet Input
+GoogleSheetsInputDialog.Shell.Title=Google Sheet Input V4
+GoogleSheetsInput.injection.SHEET=Specs for inbound Spreadsheet 
+GoogleSheetsInput.injection.spreadsheetKey=ID of the Google Spreadsheet
+GoogleSheetsInput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
+GoogleSheetsInput.injection.INPUT_FIELDS=Name of the inbound fields
+GoogleSheetsInput.injection.INPUT_NAME=Field Name
+GoogleSheetsInput.injection.INPUT_POSITION=Position
+GoogleSheetsInput.injection.INPUT_LENGTH=Length
+GoogleSheetsInput.injection.INPUT_IGNORE=Ignore
+GoogleSheetsInput.injection.INPUT_FORMAT=Format
+GoogleSheetsInput.injection.INPUT_PRECISION=Precision
+GoogleSheetsInput.injection.INPUT_CURRENCY=Currency
+GoogleSheetsInput.injection.INPUT_DECIMAL=Decimal
+GoogleSheetsInput.injection.INPUT_GROUP=Group
+GoogleSheetsInput.injection.INPUT_REPEAT=Repeat
+GoogleSheetsInput.injection.INPUT_NULL_STRING=Null String
+GoogleSheetsInput.injection.INPUT_IF_NULL=If Null
+GoogleSheetsInput.injection.jsonCrendentialPath=Path to the Google JSON credential file
+
+
+GoogleSheetsOutputDialog.TransformName.Label=Google Sheet Output
+GoogleSheetsOutputDialog.Shell.Title=Google Sheet Output V4
+GoogleSheetsOutputDialog.Create.Label=Create new sheet if it does not exist
+GoogleSheetsOutputDialog.Append.Label=Append to sheet
+GoogleSheetsOutputDialog.Share.Label=Share Full rights (RW) with
+GoogleSheetsOutputDialog.Share.LabelDW=Domain Wise Permission :
+GoogleSheetsOutput.injection.SHEET=Specs for outbound Spreadsheet
+GoogleSheetsOutput.injection.spreadsheetKey=ID of the Google Spreadsheet
+GoogleSheetsOutput.injection.worksheetId=Name of the worksheet inside the spreadsheet to write in
+GoogleSheetsOutput.injection.Create=Create Sheet if it does not exist
+GoogleSheetsOutput.injection.Append=Append data to existing Sheet (without header)
+GoogleSheetsOutput.injection.Email=Share the created sheet with a user
+GoogleSheetsOutput.injection.Domain=Domain name for domain extension
+GoogleSheetsOutput.injection.jsonCrendentialPath=Path to the Google JSON credential file
+
+
+GoogleSheetsDialog.Tab.ServiceAccount=Service Account
+GoogleSheetsDialog.PrivateKeyStore=JSON credential key file:
+GoogleSheetsDialog.Button.TestConnection=Test Connection
+GoogleSheetsDialog.Tab.Spreadsheet=Spreadsheet
+GoogleSheetsDialog.SpreadsheetKey=Spreadsheet Key
+GoogleSheetsDialog.WorksheetId=Worksheet Id
+GoogleSheetsDialog.Tab.Fields=Fields
+GoogleSheetsDialog.NrOfSampleLines=Number of sample lines to guess field types:
+
